Update build.gradle and Duke.java
The build.gradle file is outdated. In addition, there is no need
for the main and run methods in the current Duke class since we are
more focus on the GUI.

Updating build.gradle allows shadowJar to build a compatible jar
file for javaFx and updating Duke.java will improve code quality
via removable of dead code.