Add assertions to Ui.java and other Command java class files
Assertions allows duke to terminate in the case of a possible bug,
and sooner the execution stops, the safer it is.

Multiple checks for null in the parameters passed into Ui.java and
the string response returned from command execution.

By checking for null will allow us to verify and ensure that the
program is behaving as expected.