Improve code quality
Presence of dead code such as welcomeResponse() and lineResponse() in
Duke.java and unclear prominent path in certain classes.

Remove the dead code and introduce guard clauses into our program.

Removing dead code as we can simply recover it from revision control
tool when we need it in the future and use of guard clause in some code
will help us to make the code more reader friendly.