Update Junit test cases
Removing Storage test cases as it is difficult to test Storage without
proper stub and introducing additional test cases for ToDo and Unmark
command to improve unit testing of Duke.

Testing todo toStringForStorage method as well as Unmark's constructor
and isExit status.

This will ensure that the test cases closely test the program for bugs.